pyproject.toml
[project]
name = "agent-swarm"
version = "0.1.0"
description = "Multi-agent system where agents collaborate to process a user requests and generate responses"
authors = [
    {name = "pedroaires",email = "pedro.aires.cabral@gmail.com"}
]

requires-python = "^3.11"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "langgraph-supervisor (>=0.0.29,<0.0.30)",
    "tavily-python (>=0.7.11,<0.8.0)",
    "langgraph-swarm (>=0.0.14,<0.0.15)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
